sort:merge_sort remove length_of_vector
The code itself does not make any sense. sizeof(vect), where vect is
declared as: std::vector<int> vect {1, 2, 3, 4};
It does not return the size of vector but the size of this vector type

int size = vect.size(); is not safe also, with 64-bit memory system, the
size of a vector can be larger that an integer can hold.

use size_type is the safe way...