sorting: radix sort
It's always good to brush up basic sorting algo.

It brings me a question, when to use radix sort?
Radix sort time complexity: O(kn)
General sort time complexity: O(nlgn)

In order to compare which one is better, we need to compare lgn and k
where n is the total number of element to sort and k is the range of
elements

I. When k, range of elements, is large and number of element is small,
then k >> lgn, in this case, general sorting algo like merge/quick sort
prevails.
II, When k, range of elements, is small and number of element is large,
then k << lgn, say you are sorting integers with at most two digits and
n is at million level, radix sort prevails and it's O(n)